Added medical history dto
