+++ app/src/main/java/com/triadss/doctrack2/dto/AddPatientDto.java
@@ -16,14 +16,6 @@ public class AddPatientDto {
 
     private String course;
 
-    private String pastIllness;
-
-    private String prevOperation;
-
-    private String obGyneHist;
-
-    private String familyHist;
-
     public String getEmail() {
         return email;
     }
@@ -78,37 +70,5 @@ public class AddPatientDto {
 
     public void setAge(String age) {
         Age = age;
-    }
-
-    public String getPastIllness() {
-        return pastIllness;
-    }
-
-    public void setPastIllness(String pastIllness) {
-        this.pastIllness = pastIllness;
-    }
-
-    public String getPrevOperation() {
-        return prevOperation;
-    }
-
-    public void setPrevOperation(String prevOperation) {
-        this.prevOperation = prevOperation;
-    }
-
-    public String getObGyneHist() {
-        return obGyneHist;
-    }
-
-    public void setObGyneHist(String obGyneHist) {
-        this.obGyneHist = obGyneHist;
-    }
-
-    public String getFamilyHist() {
-        return familyHist;
-    }
-
-    public void setFamilyHist(String familyHist) {
-        this.familyHist = familyHist;
     }
 }
